pyproject.toml
[tool.poetry]
name = "alertist"
version = "0.1.0"
description = "A general purpose library for sending alerts"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-mock = "^3.14.0"
ruff = "^0.4.4"
pytest-asyncio = "^0.23.6"
twine = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
